requirements.txt
cmake==3.25.0
click==8.1.3
numpy==1.24.1
Pillow==9.4.0
dlib==19.24.0
face-recognition==1.3.0
opencv-python==4.6.0.66
pytz
getmac
python-dotenv==1.0.1
firebase-admin